+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - a new 'Confirm the academy name' task that collects the confirmed new academy
   name so that the Get information about schools record can be added.
 - Add hint text to External contacts tab
+- Users can view a list of projects for a given trust
 
 ### Changed